Fix checkstyle issues
