--- hudi-kafka-connect/configs/connect-sink.properties
@@ -1,22 +1,0 @@
-# Licensed to the Apache Software Foundation (ASF) under one or more
-# contributor license agreements.  See the NOTICE file distributed with
-# this work for additional information regarding copyright ownership.
-# The ASF licenses this file to You under the Apache License, Version 2.0
-# (the "License"); you may not use this file except in compliance with
-# the License.  You may obtain a copy of the License at
-#
-#    http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-name=hudi-sink
-connector.class=org.apache.hudi.connect.HudiSinkConnector
-key.converter="org.apache.kafka.connect.storage.StringConverter"
-value.converter="org.apache.kafka.connect.json.JsonConverter"
-value.converter.schemas.enable=false
-tasks.max=4
-topics=hudi-test-topic
