--- hudi-kafka-connect/src/main/java/org/apache/hudi/connect/writers/TransactionServices.java
@@ -1,15 +1,0 @@
-package org.apache.hudi.connect.writers;
-
-import org.apache.hudi.client.WriteStatus;
-
-import java.util.List;
-import java.util.Map;
-
-public interface TransactionServices {
-
-  String startCommit();
-
-  void endCommit(String commitTime, List<WriteStatus> writeStatuses, Map<String, String> extraMetadata);
-
-  Map<String, String> loadLatestCommitMetadata();
-}
